+++ src/components/ButtonBlock/Edit.jsx
@@ -3,23 +3,17 @@ import { SidebarPortal } from '@plone/volto/components';
 import ButtonSidebar from './ButtonSidebar';
 import View from './View';
 
-const Edit = ({
-  data,
-  onChangeBlock,
-  block,
-  selected,
-  sidebarData,
-  isEditMode,
-}) => {
+const Edit = (props) => {
+  const { data, block, onChangeBlock, selected } = props;
   return (
     <>
-      <View data={data} isEditMode />
+      <View {...props} isEditMode />
       <SidebarPortal selected={selected}>
         <ButtonSidebar
+          {...props}
           data={data}
           block={block}
           onChangeBlock={onChangeBlock}
-          sidebarData={sidebarData}
         />
       </SidebarPortal>
     </>