+++ src/components/ButtonBlock/View.jsx
@@ -5,21 +5,41 @@ import { ConditionalLink } from '@plone/volto/components';
 import { flattenToAppURL, isInternalURL } from '@plone/volto/helpers';
 
 const View = ({ data, isEditMode }) => {
-  let link = data?.href ? (
-    isInternalURL(data.href) ? (
-      <ConditionalLink to={flattenToAppURL(data.href)} condition={!isEditMode}>
+  const [hasLink, setHasLink] = React.useState(false);
+
+  React.useEffect(() => {
+    if (data.href) {
+      if (data.href && data.href.length > 0) {
+        setHasLink(true);
+      }
+      if (data.href.length === 0) {
+        setHasLink(false);
+      }
+    }
+
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [data.href]);
+
+  let link = hasLink ? (
+    data.href.length > 0 && isInternalURL(data.href[0]['@id']) ? (
+      <ConditionalLink
+        to={data.href.length > 0 ? flattenToAppURL(data.href[0]['@id']) : ''}
+        condition={!isEditMode}
+      >
         <Button className={(cx('button'), data.align)}>{data.title}</Button>
       </ConditionalLink>
     ) : (
-      <a href={flattenToAppURL(data.href)}>
-        <Button className={(cx('button'), data.align)}>{data.title}</Button>
-      </a>
+      data.href.length > 0 && (
+        <a href={flattenToAppURL(data.href[0]['@id'])}>
+          <Button className={(cx('button'), data.align)}>{data.title}</Button>
+        </a>
+      )
     )
   ) : (
-    <Button className={cx({ noLink: !data.href })}>{data.title}</Button>
+    <Button className="noLink">{data.title}</Button>
   );
 
-  return <div className={cx(`block button align ${data.align}`)}>{link}</div>;
+  return <div className={cx(`block button align ${data?.align}`)}>{link}</div>;
 };
 
 export default View;