+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Add id to the default server export @sneridagh
 - Export `show_navigation_portlet` field @sneridagh
 - Export with the keys ordered @sneridagh
+- Refactoring button block to use schema renderer @iFlameing
 
 ### Internal